requirements.txt
aioredis==2.0.1
elasticsearch[async]==7.9.1
fastapi==0.78.0
orjson==3.6.0
pydantic==1.9.0
uvicorn==0.14.0
elasticsearch-dsl==7.4.0
flake8==4.0
fastapi-redis-cache==0.1.0
fastapi-cache2==0.1.8
flake8==4.0
